Home shows the filter
But does not work as expected
