+++ assets/css/indexfilter.css
@@ -1,4 +1,4 @@
-.filteren button {
+.filter button {
 	display: inline-block;
 	margin: 1px 2px 1px 2px;
 	padding: 0;
